Remove useless interface
